refactor: move header data boxes to separate file
